refactor: 패키지 구조 변경 및 함수명 변경
