+++ app/src/main/java/woowacourse/shopping/presentation/productdetail/ProductDetailPresenter.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.presentation.productdetail
 
-import woowacourse.shopping.Product
 import woowacourse.shopping.data.cart.CartRepository
 import woowacourse.shopping.data.product.ProductRepository
+import woowacourse.shopping.model.Product
 import woowacourse.shopping.presentation.mapper.toPresentation
 
 class ProductDetailPresenter(
@@ -20,7 +20,7 @@ class ProductDetailPresenter(
     }
 
     override fun putProductInCart() {
-        cartRepository.addCartProductId(currentProduct.id)
+        cartRepository.addCartProduct(currentProduct.id)
         view.showCompleteMessage(currentProduct.name)
     }
 }
